# feat/tuning: optimise unnecessary allocations