--- a/internal/tree/group.go
+++ b/internal/tree/group.go
@@ -42,47 +42,43 @@ func (g *Group) Merge(stack []string, attrs ...slog.Attr) {
 func merge(in *Group, stack []string, attr slog.Attr) {
 	l := last(in, stack)
 
-	var match *int
 	for i := range l.vs {
 		if l.vs[i].name == attr.Key {
-			match = &i
-		}
-	}
+			if attr.Value.Kind() != slog.KindGroup {
+				l.vs[i].v = attr.Value
+				return
+			}
 
-	if match == nil {
-		if attr.Value.Kind() != slog.KindGroup {
-			l.vs = append(l.vs, &value{name: attr.Key, v: attr.Value})
-		} else {
-			ng := &Group{vs: make([]*value, 0)}
-			ng.Merge(nil, attr.Value.Group()...)
-			l.vs = append(l.vs, &value{name: attr.Key, g: ng})
+			if l.vs[i].g == nil {
+				l.vs[i].g = new(Group)
+			}
+
+			l.vs[i].g.Merge(nil, attr.Value.Group()...)
+			return
 		}
-		return
 	}
 
 	if attr.Value.Kind() != slog.KindGroup {
-		l.vs[*match] = &value{name: attr.Key, v: attr.Value}
-		return
+		l.vs = append(l.vs, &value{name: attr.Key, v: attr.Value})
+	} else {
+		ng := new(Group)
+		ng.Merge(nil, attr.Value.Group()...)
+		l.vs = append(l.vs, &value{name: attr.Key, g: ng})
 	}
 
-	if l.vs[*match].g == nil {
-		ng := &Group{vs: make([]*value, 0)}
-		l.vs[*match] = &value{name: attr.Key, g: ng} // override
-	}
-
-	l.vs[*match].g.Merge(nil, attr.Value.Group()...)
+	return
 }
 
 func last(in *Group, stack []string) *Group {
-	switch len(stack) {
-	case 0:
+	if len(stack) == 0 {
 		return in
-	default:
-		for _, v := range in.vs {
-			if v.name == stack[0] {
-				return last(v.g, stack[1:])
-			}
-		}
-		panic("slog/meld: invalid or misconfigured stack")
 	}
+
+	for _, v := range in.vs {
+		if v.name == stack[0] {
+			return last(v.g, stack[1:])
+		}
+	}
+
+	panic("slog/meld: invalid or misconfigured stack")
 }