`benchstat` diff (n=64):

```
$ benchstat before.bench after.bench
goos: linux
goarch: amd64
pkg: snqk.dev/slog/meld
cpu: Intel(R) Core(TM) i7-5960X CPU @ 3.00GHz
              │ before.bench │             after.bench             │
              │    sec/op    │    sec/op     vs base               │
MeldLogger      6.771µ ±  3%   6.022µ ±  4%  -11.07% (p=0.000 n=8)
MeldLogger-2    3.961µ ±  2%   3.430µ ±  4%  -13.41% (p=0.000 n=8)
MeldLogger-4    2.156µ ±  9%   1.957µ ±  6%   -9.23% (p=0.002 n=8)
MeldLogger-8    2.646µ ±  8%   2.118µ ± 13%  -19.94% (p=0.000 n=8)
MeldLogger-16   3.497µ ± 13%   2.492µ ±  3%  -28.74% (p=0.000 n=8)
geomean         3.513µ         2.923µ        -16.80%

              │ before.bench │             after.bench             │
              │     B/op     │     B/op      vs base               │
MeldLogger      3.141Ki ± 0%   2.336Ki ± 0%  -25.62% (p=0.000 n=8)
MeldLogger-2    3.141Ki ± 0%   2.336Ki ± 0%  -25.62% (p=0.000 n=8)
MeldLogger-4    3.143Ki ± 0%   2.337Ki ± 0%  -25.64% (p=0.000 n=8)
MeldLogger-8    3.147Ki ± 0%   2.340Ki ± 0%  -25.65% (p=0.000 n=8)
MeldLogger-16   3.155Ki ± 0%   2.347Ki ± 0%  -25.62% (p=0.000 n=8)
geomean         3.145Ki        2.339Ki       -25.63%

              │ before.bench │            after.bench            │
              │  allocs/op   │ allocs/op   vs base               │
MeldLogger        73.00 ± 0%   62.00 ± 0%  -15.07% (p=0.000 n=8)
MeldLogger-2      73.00 ± 0%   62.00 ± 0%  -15.07% (p=0.000 n=8)
MeldLogger-4      73.00 ± 0%   62.00 ± 0%  -15.07% (p=0.000 n=8)
MeldLogger-8      73.00 ± 0%   62.00 ± 0%  -15.07% (p=0.000 n=8)
MeldLogger-16     73.00 ± 0%   62.00 ± 0%  -15.07% (p=0.000 n=8)
geomean           73.00        62.00       -15.07%
```